Remove unused code and old data stores
