+++ utils.go
@@ -92,30 +92,6 @@ func JsonUnmarshal(s string) (map[string]interface{}, error) {
 	return dat, nil
 }
 
-// GetStoreURLFromDsHref extracts the base store url from the href provied in the DATASOURCE_[name] environment variable.
-func GetStoreURLFromDsHref(href string) (string, error) {
-
-	u, err := url.Parse(href)
-	if err != nil {
-		return "", err
-	}
-
-	return u.Scheme + "://" + u.Host, nil
-
-}
-
-// GetDsIdFromDsHref extracts the base data source ID from the href provied in the DATASOURCE_[name] environment variable.
-func GetDsIdFromDsHref(href string) (string, error) {
-
-	u, err := url.Parse(href)
-	if err != nil {
-		return "", err
-	}
-
-	return s.Replace(u.Path, "/", "", -1), nil
-
-}
-
 func makeArbiterRequest(arbMethod string, path string, hostname string, endpoint string, method string) (string, int) {
 
 	var jsonStr = []byte(`{"target":"` + hostname + `","path":"` + endpoint + `","method":"` + method + `"}`)
@@ -179,152 +155,6 @@ func checkTokenCache(href string, method string) (string, error) {
 	return tokenCache[routeHash], nil
 }
 
-func makeStoreRequest(href string, method string) (string, error) {
-
-	method = s.ToUpper(method)
-	token, err := checkTokenCache(href, method)
-	if err != nil {
-		return "", err
-	}
-
-	//perform store request with token
-	req, err := http.NewRequest(method, href, nil)
-	req.Header.Set("X-Api-Key", token)
-	req.Header.Set("Content-Type", "application/json")
-	req.Close = true
-	resp, err := databoxClient.Do(req)
-	if err != nil {
-		panic(err)
-	}
-	defer resp.Body.Close()
-
-	body, err1 := ioutil.ReadAll(resp.Body)
-	if err1 != nil {
-		return "", err1
-	}
-
-	if resp.StatusCode < 200 || resp.StatusCode >= 300 {
-		return "", errors.New(string(body[:]))
-	}
-
-	return string(body[:]), nil
-}
-
-func makeStoreRequestPOST(href string, data string) (string, error) {
-
-	method := "POST"
-	token, err := checkTokenCache(href, method)
-	if err != nil {
-		return "", err
-	}
-
-	//perform store request with token
-	req, err := http.NewRequest(method, href, bytes.NewBufferString(data))
-	req.Header.Set("X-Api-Key", token)
-	req.Header.Set("Content-Type", "application/json")
-	req.Close = true
-	resp, err := databoxClient.Do(req)
-	if err != nil {
-		return "", err
-	}
-	defer resp.Body.Close()
-
-	body, err1 := ioutil.ReadAll(resp.Body)
-	if err1 != nil {
-		return "", err1
-	}
-
-	return string(body[:]), nil
-}
-
-func makeStoreRequestForm(href string, method string, values map[string]string) (string, error) {
-
-	method = s.ToUpper(method)
-	token, err := checkTokenCache(href, method)
-	if err != nil {
-		return "", err
-	}
-
-	// form-style parameters
-	data := url.Values{}
-	for k, v := range values {
-		data.Set(k, v)
-	}
-	encdata := data.Encode()
-	fmt.Printf("form request %s %s %s\n", href, method, encdata)
-	//perform store request with token
-	req, err := http.NewRequest(method, href, bytes.NewBufferString(encdata))
-	req.Header.Set("X-Api-Key", token)
-	req.Header.Set("Content-Type", "application/x-www-form-urlencoded")
-	req.Close = true
-	resp, err := databoxClient.Do(req)
-	if err != nil {
-		return "", err
-	}
-	defer resp.Body.Close()
-
-	body, err1 := ioutil.ReadAll(resp.Body)
-	if err1 != nil {
-		return "", err1
-	}
-
-	return string(body[:]), nil
-}
-
-func makeStoreRequestJson(href string, method string, data string) (string, error) {
-
-	method = s.ToUpper(method)
-	token, err := checkTokenCache(href, method)
-	if err != nil {
-		return "", err
-	}
-
-	//perform store request with token
-	req, err := http.NewRequest(method, href, bytes.NewBufferString(data))
-	req.Header.Set("X-Api-Key", token)
-	req.Header.Set("Content-Type", "application/json")
-	req.Close = true
-	resp, err := databoxClient.Do(req)
-	if err != nil {
-		return "", err
-	}
-	defer resp.Body.Close()
-
-	body, err1 := ioutil.ReadAll(resp.Body)
-	if err1 != nil {
-		return "", err1
-	}
-
-	if resp.StatusCode < 200 || resp.StatusCode >= 300 {
-		return "", errors.New(string(body[:]))
-	}
-
-	return string(body[:]), nil
-}
-
-//WaitForStoreStatus will wait for the store available at href to respond with an active status.
-func WaitForStoreStatus(href string) {
-
-	h, err := GetStoreURLFromDsHref(href)
-
-	if err == nil {
-		resp, err1 := databoxClient.Get(h + "/status")
-		if err1 == nil {
-			defer resp.Body.Close()
-			_, err = ioutil.ReadAll(resp.Body)
-		} else {
-			err = err1
-		}
-	}
-
-	if err != nil {
-		fmt.Printf("[waitForStoreStatus] Retrying in 1s...")
-		time.Sleep(1000 * time.Millisecond)
-		WaitForStoreStatus(href)
-	}
-
-}
-
 type StoreMetadata struct {
 	Description    string
 	ContentType    string
@@ -350,52 +180,4 @@ type relValPairBool struct {
 type hypercat struct {
 	ItemMetadata []interface{} `json:"item-metadata"`
 	Href         string        `json:"href"`
-}
-
-// RegisterDatasource is used by apps and drivers to register datasource in stores they
-// own.
-func RegisterDatasource(href string, metadata StoreMetadata) (string, error) {
-
-	catURL, _ := GetStoreURLFromDsHref(href)
-	catURL = catURL + "/cat"
-
-	if metadata.Description == "" ||
-		metadata.ContentType == "" ||
-		metadata.Vendor == "" ||
-		metadata.DataSourceType == "" ||
-		metadata.DataSourceID == "" ||
-		metadata.StoreType == "" {
-
-		return "", errors.New("Missing required metadata")
-
-	}
-
-	cat := hypercat{}
-	cat.ItemMetadata = append(cat.ItemMetadata, relValPair{Rel: "urn:X-hypercat:rels:hasDescription:en", Val: metadata.Description})
-	cat.ItemMetadata = append(cat.ItemMetadata, relValPair{Rel: "urn:X-hypercat:rels:isContentType", Val: metadata.ContentType})
-	cat.ItemMetadata = append(cat.ItemMetadata, relValPair{Rel: "urn:X-databox:rels:hasVendor", Val: metadata.Vendor})
-	cat.ItemMetadata = append(cat.ItemMetadata, relValPair{Rel: "urn:X-databox:rels:hasType", Val: metadata.DataSourceType})
-	cat.ItemMetadata = append(cat.ItemMetadata, relValPair{Rel: "urn:X-databox:rels:hasDatasourceid", Val: metadata.DataSourceID})
-	cat.ItemMetadata = append(cat.ItemMetadata, relValPair{Rel: "urn:X-databox:rels:hasStoreType", Val: metadata.StoreType})
-
-	if metadata.IsActuator {
-		cat.ItemMetadata = append(cat.ItemMetadata, relValPairBool{Rel: "urn:X-databox:rels:isActuator", Val: true})
-	}
-
-	if metadata.Location != "" {
-		cat.ItemMetadata = append(cat.ItemMetadata, relValPair{Rel: "urn:X-databox:rels:hasLocation", Val: metadata.Location})
-	}
-
-	if metadata.Unit != "" {
-		cat.ItemMetadata = append(cat.ItemMetadata, relValPair{Rel: "urn:X-databox:rels:hasUnit", Val: metadata.Unit})
-	}
-
-	href, _ = GetStoreURLFromDsHref(href)
-	cat.Href = href + "/" + metadata.DataSourceID
-
-	jsonByteArray, _ := json.Marshal(cat)
-
-	fmt.Println(string(jsonByteArray[:]))
-
-	return makeStoreRequestPOST(catURL, string(jsonByteArray[:]))
 }
